Major refactor.
