+++ kavaplugin/src/main/java/com/kaltura/playkit/plugins/kava/KavaMediaEntryType.java
@@ -1,0 +1,12 @@
+package com.kaltura.playkit.plugins.kava;
+
+/**
+ * Created by anton.afanasiev on 18/02/2018.
+ */
+
+enum KavaMediaEntryType {
+    Vod,
+    Live,
+    Dvr,
+    Unknown
+}
